refactor: 将主键类型从 Long 更改为 Integer，优化数据模型和数据库访问层

+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/ItemAttributeRelationRepository.java
@@ -13,5 +13,5 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface ItemAttributeRelationRepository
-    extends ListCrudRepository<ItemAttributeRelation, Long>,
-        ListPagingAndSortingRepository<ItemAttributeRelation, Long> {}
+    extends ListCrudRepository<ItemAttributeRelation, Integer>,
+        ListPagingAndSortingRepository<ItemAttributeRelation, Integer> {}
